feat: improve template code
