+++ scripts/cli/__init__.py
@@ -1,1 +1,3 @@
 """Create configured argparse's parser."""
+
+__version__ = "1.0.0"  # major.minor.patch
