+++ scripts/example/__init__.py
@@ -1,1 +1,3 @@
 """cli package example of use."""
+
+__version__ = "1.0.0"  # major.minor.patch
